Rewrite MoreOrLess in TS
