--- src/MoreOrLess/index.d.ts
@@ -1,2 +1,0 @@
-import MoreOrLess from './MoreOrLess';
-export default MoreOrLess;
