+++ src/MoreOrLess/index.ts
@@ -1,0 +1,3 @@
+import MoreOrLess, { MoreOrLessProps } from './MoreOrLess';
+export default MoreOrLess;
+export { MoreOrLessProps };
